/libs/*: libraries required for knime nodes. classpath: updated accordingly. settings: consistenly use * for imports. RegressionMeasure/ RegressionMeasureCooki/ Table: break for case. DAGView/ QualityMeasure/ CortanaNodeView/ ResultWindow: removed redundant cast. linux_compile.sh / linux_jar.sh: updated for new libs. NOTE windows compile scripts not updated, needs windows. NOTE all other classes now consistently use * for imports.
git-svn-id: https://svn.science.uu.nl/repos/sci.knobb101.SubgroupDiscovery@1336 3a85a776-1598-11dc-ad04-adf9e166afdd

+++ SubDisc/src/nl/liacs/subdisc/QualityMeasure.java
@@ -452,7 +452,7 @@ public class QualityMeasure
 				float aComplementCoverage = itsNrRecords - theCoverage;
 				float aMean = (theCoverage*(theCoverage+aComplementCoverage+1))/2.0f;
 				float aStDev = (float) Math.sqrt((theCoverage*aComplementCoverage*(theCoverage+aComplementCoverage+1))/12.0f);
-				aReturn = (float) Math.abs((theSum-aMean)/aStDev); break;
+				aReturn = Math.abs((theSum-aMean)/aStDev); break;
 			}
 			case MMAD : { aReturn = (theCoverage/(2*theMedian+theMedianAD)); break; }
 		}
